requirements.txt
streamlit
pandas
langchain
langchain-core
langchain-google-genai
pydantic
imap-tools
python-dotenv
google-generativeai
